# docs/swagger.json
{"swagger":"2.0","info":{"title":"yildiz HTTP interface","description":"Graph Database on top of MySQL","version":"1.18.0"},"host":"localhost","schemes":["http"],"consumes":["application/json"],"produces":["application/json"],"paths":{"/documentation":{"get":{"description":"Documentation route, if yaml is true the payload will be in yaml format","parameters":[{"type":"boolean","name":"yaml","in":"query"}],"responses":{"200":{"schema":{"description":"The swagger definition of the routes","type":"object","additionalProperties":true},"description":"The swagger definition of the routes"}}}},"/":{"get":{"responses":{"2xx":{"schema":{"type":"object","properties":{"version":{"type":"string"}},"additionalProperties":false},"description":"Default Response"}}}},"/admin/health":{"get":{"responses":{"2xx":{"schema":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":false},"description":"Default Response"}}}},"/admin/healthcheck":{"get":{"responses":{"2xx":{"schema":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":false},"description":"Default Response"}}}},"/admin/stats":{"get":{"responses":{"2xx":{"schema":{"type":"object","additionalProperties":true},"description":"Default Response"}}}},"/admin/metrics":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/translator/{identifier}":{"get":{"parameters":[{"name":"identifier","in":"path","required":true,"type":"string"},{"name":"required","in":"path","required":true}],"responses":{"2xx":{"schema":{"type":"object","properties":{"identifier":{"type":"number"},"value":{"type":"string"},"data":{"type":"object","additionalProperties":true},"ttld":{"type":"boolean"},"created_at":{"type":"string"}},"additionalProperties":false},"description":"Default Response"}}},"delete":{"parameters":[{"name":"identifier","in":"path","required":true,"type":"string"},{"name":"required","in":"path","required":true}],"responses":{"2xx":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}},"additionalProperties":false},"description":"Default Response"}}}},"/translator/translate-and-store":{"post":{"responses":{"2xx":{"schema":{"type":"object","properties":{"identifier":{"type":"number"},"value":{"type":"string"},"data":{"type":"object","additionalProperties":true},"ttld":{"type":"boolean"},"created_at":{"type":"string"}},"additionalProperties":false},"description":"Default Response"}}}},"/path/shortest-path":{"post":{"parameters":[{"name":"body","in":"body","schema":{"start":{"type":"number"},"end":{"type":"number"},"required":["start","end"]}}],"responses":{"2xx":{"schema":{"type":"object","properties":{"paths":{"type":"array","items":{"type":"object","additionalProperties":true}}},"additionalProperties":false},"description":"Default Response"}}}},"/node/{identifier}":{"get":{"parameters":[{"name":"identifier","in":"path","required":true,"type":"string"},{"name":"required","in":"path","required":true}],"responses":{"2xx":{"schema":{"type":"object","properties":{"id":{"type":"number"},"identifier":{"type":"string"},"data":{"type":"object","additionalProperties":true},"ttld":{"type":"boolean"},"created_at":{"type":"string"}},"additionalProperties":false},"description":"Default Response"}}},"delete":{"parameters":[{"name":"identifier","in":"path","required":true,"type":"string"},{"name":"required","in":"path","required":true}],"responses":{"2xx":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}},"additionalProperties":false},"description":"Default Response"}}}},"/node":{"post":{"responses":{"2xx":{"schema":{"type":"object","properties":{"id":{"type":"number"},"identifier":{"type":"string"},"data":{"type":"object","additionalProperties":true},"ttld":{"type":"boolean"},"created_at":{"type":"string"}},"additionalProperties":false},"description":"Default Response"}}}},"/edge/{leftId}/{rightId}/{relation}":{"get":{"parameters":[{"name":"leftId","in":"path","required":true,"type":"number"},{"name":"rightId","in":"path","required":true,"type":"number"},{"name":"relation","in":"path","required":true,"type":"string"},{"name":"required","in":"path","required":true}],"responses":{"2xx":{"schema":{"type":"object","properties":{"id":{"type":"number"},"depth":{"type":"number"},"relation":{"type":"string"},"data":{"type":"object","additionalProperties":true},"ttld":{"type":"boolean"},"created_at":{"type":"string"}},"additionalProperties":false},"description":"Default Response"}}},"delete":{"parameters":[{"name":"leftId","in":"path","required":true,"type":"number"},{"name":"rightId","in":"path","required":true,"type":"number"},{"name":"relation","in":"path","required":true,"type":"string"},{"name":"required","in":"path","required":true}],"responses":{"2xx":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}},"additionalProperties":false},"description":"Default Response"}}}},"/edge":{"post":{"responses":{"2xx":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}},"additionalProperties":false},"description":"Default Response"}}}},"/edge/depth/increase":{"put":{"responses":{"2xx":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}},"additionalProperties":false},"description":"Default Response"}}}},"/edge/depth/decrease":{"put":{"responses":{"2xx":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}},"additionalProperties":false},"description":"Default Response"}}}},"/edge/left/{id}/{relation}":{"get":{"parameters":[{"name":"id","in":"path","required":true},{"name":"relation","in":"path","required":true,"type":"string"},{"name":"required","in":"path","required":true}],"responses":{"2xx":{"schema":{"type":"object","properties":{"edges":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"depth":{"type":"number"},"relation":{"type":"string"},"data":{"type":"object","additionalProperties":true},"ttld":{"type":"boolean"},"created_at":{"type":"string"},"right_node_id":{"type":"number"}}}}},"additionalProperties":false},"description":"Default Response"}}}},"/edge/right/{id}/{relation}":{"get":{"parameters":[{"name":"id","in":"path","required":true},{"name":"relation","in":"path","required":true,"type":"string"},{"name":"required","in":"path","required":true}],"responses":{"2xx":{"schema":{"type":"object","properties":{"edges":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"depth":{"type":"number"},"relation":{"type":"string"},"data":{"type":"object","additionalProperties":true},"ttld":{"type":"boolean"},"created_at":{"type":"string"},"left_node_id":{"type":"number"}}}}},"additionalProperties":false},"description":"Default Response"}}}},"/edge/both/{id}/{relation}":{"get":{"parameters":[{"name":"id","in":"path","required":true},{"name":"relation","in":"path","required":true,"type":"string"},{"name":"required","in":"path","required":true}],"responses":{"2xx":{"schema":{"type":"object","properties":{"edges":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"depth":{"type":"number"},"relation":{"type":"string"},"data":{"type":"object","additionalProperties":true},"ttld":{"type":"boolean"},"created_at":{"type":"string"},"right_node_id":{"type":"number"},"left_node_id":{"type":"number"}}}}},"additionalProperties":false},"description":"Default Response"}}}},"/access/translated-edge-info":{"post":{"parameters":[{"name":"body","in":"body","schema":{"values":{"type":"array","items":{"type":"string"}},"required":["values"]}}],"responses":{"2xx":{"schema":{"type":"object","properties":{"edges":{"type":"array","items":{"type":"object","properties":{"other_node_id":{"type":"number"},"relation":{"type":"string"},"depth":{"type":"number"},"edata":{"type":"object","additionalProperties":true},"value":{"type":"string"},"tdata":{"type":"object","additionalProperties":true}}}}},"additionalProperties":false},"description":"Default Response"}}}}}}